create components => layout => (header), some fixations in (Header.vue, Menu.vue, Column.vue, Footer.vue) components
