--- components/layout/Header.vue
@@ -1,23 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-	<div class="fixed top-0 left-0 min-h-[120px] max-h-[120px] bg-black w-full">
-		<div class="flex items-cetner min-h-[120px] max-h-[120px] px-[85px] justify-between ">
-			<div class="flex items-center text-[26px] text-white">
-				<Icon name="iconamoon:search-thin" />
-			</div>
-
-			<LayoutMenu/>
-
-			<div class="flex items-center text-white">
-				<Icon   name="iconamoon:arrow-up-2-light" size="18" />
-			</div>
-		</div>
-	</div>
-</template>
-
-<style scoped>
-
-</style>
