+++ layouts/default.vue
@@ -3,14 +3,14 @@
 <template>
 
 
-<section class="relative w-full h-[100vh]">
-    <LayoutHeader />
-    <div class="w-full mt-[80px]">
-      <slot />
-    </div>
-    <LayoutFooter />
+	<section class="relative w-full h-[100vh]">
+		<LayoutHeader />
+		<div class="w-full mt-[80px]">
+			<slot />
+		</div>
+		<LayoutFooter />
 
-  </section>
-  
+	</section>
+
 </template>
 <style scoped></style>
